Pastel de Flango

+++ agenciaturismo/src/main/java/br/senac/tads3/pi03b/gruposete/models/Voo.java
@@ -73,7 +73,7 @@ public class Voo {
         this.quantidade_passagens = quantidade_passagens;
     }
 
-    public double getPreco() {
+    public float getPreco() {
         return preco;
     }
